+++ imager/src/exfat.rs
@@ -1,10 +1,10 @@
 use std::{
     fmt,
     fs,
-    mem,
     path,
 };
 
+mod boot_checksum_sector;
 mod boot_sector;
 mod extended_boot_sector;
 mod oem_parameter_sector;
@@ -16,7 +16,7 @@ pub struct Exfat {
     extended_boot_sectors: [extended_boot_sector::ExtendedBootSector; 0x8],
     oem_parameter_sector: oem_parameter_sector::OemParameterSector,
     reserved_sector: reserved_sector::ReservedSector,
-    boot_checksum_sector: Option<BootChecksumSector>,
+    boot_checksum_sector: Option<boot_checksum_sector::BootChecksumSector>,
 }
 
 impl Exfat {
@@ -32,7 +32,7 @@ impl Exfat {
     }
 
     fn checksum(self) -> Self {
-        let boot_checksum_sector = BootChecksumSector::new(&self);
+        let boot_checksum_sector = boot_checksum_sector::BootChecksumSector::new(&self);
         Self {
             boot_sector: self.boot_sector,
             extended_boot_sectors: self.extended_boot_sectors,
@@ -105,54 +105,5 @@ trait Packable {
 trait Unpackable {
     type Unpacked;
     fn unpack(&self) -> Self::Unpacked;
-}
-
-#[derive(Clone, Copy, Debug)]
-struct BootChecksumSector {
-    checksum: [u32; mem::size_of::<RawSector>() / mem::size_of::<u32>()],
-}
-
-impl BootChecksumSector {
-    fn new(exfat: &Exfat) -> Self {
-        let mut sectors: Vec<Box<dyn Sector>> = vec![];
-        sectors.push(Box::new(exfat.boot_sector));
-        for extended_boot_sector in exfat.extended_boot_sectors {
-            sectors.push(Box::new(extended_boot_sector));
-        }
-        sectors.push(Box::new(exfat.oem_parameter_sector));
-        sectors.push(Box::new(exfat.reserved_sector));
-        let checksum: u32 = sectors
-            .into_iter()
-            .map(|sector| sector.to_bytes().to_vec())
-            .flatten()
-            .enumerate()
-            .filter(|(i, _)| match i {
-                106 | 107 | 112 => false,
-                _ => true,
-            })
-            .map(|(_, byte)| byte)
-            .fold(0 as u32, |checksum, byte| match checksum & 1 {
-                1 => 0x80000000,
-                0 => 0,
-                _ => panic!("Can't create checksum sector."),
-            } + (checksum >> 1) + (byte as u32));
-        Self {
-            checksum: [checksum; 0x80],
-        }
-    }
-}
-
-impl Sector for BootChecksumSector {
-    fn to_bytes(&self) -> RawSector {
-        unsafe {
-            mem::transmute::<Self, RawSector>(*self)
-        }
-    }
-}
-
-impl fmt::Display for BootChecksumSector {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "boot_checksum_sector.checksum = {:x?}", self.checksum)
-    }
 }
